ISSUE #3657 - Refactoring of selects
