+++ frontend/stories/inputs/select/SelectWithLabel.stories.tsx
@@ -14,7 +14,7 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import { SelectWithLabel } from '@controls/searchSelect/searchSelect.component';
+import { SelectWithLabel } from '@controls/selectWithLabel/selectWithLabel.component';
 import { MenuItem, SelectChangeEvent } from '@mui/material';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { useState } from 'react';
